master.yaml uses JSONPointer insteaed of JSONPath
`op`, `path`, `value` are the fields needed to parse JSONPointer, rest
of the fields are just for informative purpose.

possible values for `op` (operation) are `"add"`, `"remove"`, `"replace"`, `"move"`, `"test"` and `"copy"`.

Assumptions:
- It is assumed in JSONPointer that `ansibletest` container is the first
one in YAML when it is parsed.
- Parameters represented in current `master.yaml` is not already present
in the litmus test spec. (since I have used `add` in `op` field of each
param)

Note: All the above assumptions are met in this commit

Signed-off-by: Abhishek Kashyap <abhishek.kashyap@mayadata.io>